volume dependency

+++ console/repositories/app_config.py
@@ -12,10 +12,8 @@ from console.utils.shortcuts import get_object_or_404
 from django.db.models import Q
 from www.db.base import BaseConnection
 from www.models.main import (GatewayCustomConfiguration, ImageServiceRelation, ServiceAttachInfo, ServiceCreateStep,
-                             ServiceDomain, ServiceDomainCertificate, ServicePaymentNotify, ServiceTcpDomain,
-                             TenantServiceAuth,
-                             TenantServiceConfigurationFile, TenantServiceEnv, TenantServiceEnvVar,
-                             TenantServiceMountRelation,
+                             ServiceDomain, ServiceDomainCertificate, ServicePaymentNotify, ServiceTcpDomain, TenantServiceAuth,
+                             TenantServiceConfigurationFile, TenantServiceEnv, TenantServiceEnvVar, TenantServiceMountRelation,
                              TenantServiceRelation, TenantServicesPort, TenantServiceVolume, ThirdPartyServiceEndpoints)
 from www.models.service_publish import ServiceExtendMethod
 
@@ -30,8 +28,7 @@ class TenantServiceEnvVarRepository(object):
         return TenantServiceEnvVar.objects.filter(tenant_id=tenant_id, service_id=service_id, scope=scope).all()
 
     def get_by_attr_name_and_scope(self, tenant_id, service_id, attr_name, scope):
-        envs = TenantServiceEnvVar.objects.filter(tenant_id=tenant_id, service_id=service_id, attr_name=attr_name,
-                                                  scope=scope)
+        envs = TenantServiceEnvVar.objects.filter(tenant_id=tenant_id, service_id=service_id, attr_name=attr_name, scope=scope)
         if envs:
             return envs[0]
         return None
@@ -52,8 +49,7 @@ class TenantServiceEnvVarRepository(object):
             ID=env_id)
 
     def get_env_by_ids_and_attr_names(self, tenant_id, service_ids, attr_names):
-        envs = TenantServiceEnvVar.objects.filter(tenant_id=tenant_id, service_id__in=service_ids,
-                                                  attr_name__in=attr_names)
+        envs = TenantServiceEnvVar.objects.filter(tenant_id=tenant_id, service_id__in=service_ids, attr_name__in=attr_names)
         return envs
 
     def get_depend_outer_envs_by_ids(self, tenant_id, service_ids):
@@ -91,8 +87,7 @@ class TenantServiceEnvVarRepository(object):
         TenantServiceEnvVar.objects.filter(pk=pk).delete()
 
     def delete_service_env_by_port(self, tenant_id, service_id, container_port):
-        TenantServiceEnvVar.objects.filter(tenant_id=tenant_id, service_id=service_id,
-                                           container_port=container_port).delete()
+        TenantServiceEnvVar.objects.filter(tenant_id=tenant_id, service_id=service_id, container_port=container_port).delete()
 
     def update_env_var(self, tenant_id, service_id, attr_name, **update_params):
         TenantServiceEnvVar.objects.filter(
@@ -103,8 +98,7 @@ class TenantServiceEnvVarRepository(object):
         default_envs = Q(attr_name__in=("COMPILE_ENV", "NO_CACHE", "DEBUG", "PROXY", "SBT_EXTRAS_OPTS"))
         prefix_start_env = Q(attr_name__startswith="BUILD_")
         build_start_env = Q(scope="build")
-        build_envs = self.get_service_env(tenant_id, service_id).filter(
-            default_envs | prefix_start_env | build_start_env)
+        build_envs = self.get_service_env(tenant_id, service_id).filter(default_envs | prefix_start_env | build_start_env)
         for benv in build_envs:
             attr_name = benv.attr_name
             if attr_name.startswith("BUILD_"):
@@ -154,8 +148,7 @@ class TenantServicePortRepository(object):
         return TenantServicesPort.objects.filter(tenant_id=tenant_id, service_id=service_id)
 
     def get_service_port_by_port(self, tenant_id, service_id, container_port):
-        ports = TenantServicesPort.objects.filter(tenant_id=tenant_id, service_id=service_id,
-                                                  container_port=container_port)
+        ports = TenantServicesPort.objects.filter(tenant_id=tenant_id, service_id=service_id, container_port=container_port)
         if ports:
             return ports[0]
         return None
@@ -168,8 +161,7 @@ class TenantServicePortRepository(object):
         TenantServicesPort.objects.filter(tenant_id=tenant_id, service_id=service_id).delete()
 
     def delete_serivce_port_by_port(self, tenant_id, service_id, container_port):
-        TenantServicesPort.objects.filter(tenant_id=tenant_id, service_id=service_id,
-                                          container_port=container_port).delete()
+        TenantServicesPort.objects.filter(tenant_id=tenant_id, service_id=service_id, container_port=container_port).delete()
 
     def delete_service_port_by_pk(self, pk):
         TenantServicesPort.objects.filter(pk=pk).delete()
@@ -239,8 +231,7 @@ class TenantServicePortRepository(object):
 class TenantServiceVolumnRepository(object):
     def list_custom_volumes(self, service_ids):
         return TenantServiceVolume.objects.filter(service_id__in=service_ids).exclude(
-            volume_type__in=["config-file", TenantServiceVolume.SHARE, TenantServiceVolume.LOCAL,
-                             TenantServiceVolume.TMPFS])
+            volume_type__in=["config-file", TenantServiceVolume.SHARE, TenantServiceVolume.LOCAL, TenantServiceVolume.TMPFS])
 
     def get_service_volumes_with_config_file(self, service_id):
         return TenantServiceVolume.objects.filter(service_id=service_id)
@@ -275,8 +266,9 @@ class TenantServiceVolumnRepository(object):
     def delete_volume_by_id(self, volume_id):
         TenantServiceVolume.objects.filter(ID=volume_id).delete()
 
-    def delete_file_by_volume_id(self, volume_id):
-        TenantServiceConfigurationFile.objects.filter(volume_id=volume_id).delete()
+    @staticmethod
+    def delete_file_by_volume(volume: TenantServiceVolume):
+        TenantServiceConfigurationFile.objects.filter(Q(volume_id=volume.ID) | Q(volume_name=volume.volume_name)).delete()
 
     def add_service_config_file(self, **service_config_file):
         return TenantServiceConfigurationFile.objects.create(**service_config_file)
@@ -284,8 +276,9 @@ class TenantServiceVolumnRepository(object):
     def get_service_config_files(self, service_id):
         return TenantServiceConfigurationFile.objects.filter(service_id=service_id)
 
-    def get_service_config_file(self, volume_id):
-        return TenantServiceConfigurationFile.objects.filter(volume_id=volume_id).first()
+    @staticmethod
+    def get_service_config_file(volume: TenantServiceVolume):
+        return TenantServiceConfigurationFile.objects.filter(Q(volume_id=volume.ID) | Q(volume_name=volume.volume_name)).first()
 
     def get_services_volumes(self, service_ids):
         return TenantServiceVolume.objects.filter(service_id__in=service_ids)
@@ -298,6 +291,10 @@ class TenantServiceVolumnRepository(object):
 
     def delete_config_files(self, sid):
         TenantServiceConfigurationFile.objects.filter(service_id=sid).defer()
+
+    @staticmethod
+    def bulk_create(volumes):
+        TenantServiceVolume.objects.bulk_create(volumes)
 
     def bulk_create_or_update(self, volumes):
         for volume in volumes:
@@ -313,8 +310,21 @@ class TenantServiceVolumnRepository(object):
         volume.save()
 
 
+class ComponentConfigurationFileRepository(object):
+    @staticmethod
+    def bulk_create(config_files):
+        TenantServiceConfigurationFile.objects.bulk_create(config_files)
+
+    @staticmethod
+    def bulk_create_or_update(config_files):
+        config_file_ids = [cf.ID for cf in config_files]
+        TenantServiceConfigurationFile.objects.filter(pk__in=config_file_ids).delete()
+        TenantServiceConfigurationFile.objects.bulk_create(config_files)
+
+
 def bulk_create_or_update(tenant_id, component_deps):
-    TenantServiceRelation.objects.filter(tenant_id=tenant_id, service_id__in=[dep.service_id for dep in component_deps]).delete()
+    TenantServiceRelation.objects.filter(
+        tenant_id=tenant_id, service_id__in=[dep.service_id for dep in component_deps]).delete()
     TenantServiceRelation.objects.bulk_create(component_deps)
 
 
@@ -323,8 +333,7 @@ class TenantServiceRelationRepository(object):
         return TenantServiceRelation.objects.filter(tenant_id=tenant_id, service_id=service_id)
 
     def get_depency_by_serivce_id_and_dep_service_id(self, tenant_id, service_id, dep_service_id):
-        deps = TenantServiceRelation.objects.filter(tenant_id=tenant_id, service_id=service_id,
-                                                    dep_service_id=dep_service_id)
+        deps = TenantServiceRelation.objects.filter(tenant_id=tenant_id, service_id=service_id, dep_service_id=dep_service_id)
         if deps:
             return deps[0]
         return None
@@ -450,9 +459,7 @@ class TenantServiceMntRelationRepository(object):
         return dep_mnts
 
     def list_mnt_relations_by_service_ids(self, tenant_id, service_ids):
-        dep_mnts = TenantServiceMountRelation.objects.filter(tenant_id=tenant_id, service_id__in=service_ids)
-
-        return dep_mnts
+        return TenantServiceMountRelation.objects.filter(tenant_id=tenant_id, service_id__in=service_ids)
 
     def add_service_mnt_relation(self, tenant_id, service_id, dep_service_id, mnt_name, mnt_dir):
         tsr = TenantServiceMountRelation.objects.create(
@@ -477,6 +484,10 @@ class TenantServiceMntRelationRepository(object):
 
     def bulk_create(self, mnts):
         TenantServiceMountRelation.objects.bulk_create(mnts)
+
+    def bulk_create_or_update(self, tenant_id, volume_deps):
+        TenantServiceMountRelation.objects.filter(tenant_id=tenant_id, pk__in=[dep.ID for dep in volume_deps]).delete()
+        self.bulk_create(volume_deps)
 
 
 class ImageServiceRelationRepository(object):
@@ -547,8 +558,7 @@ class ServiceDomainRepository(object):
         except ServiceDomain.DoesNotExist:
             return None
 
-    def get_domain_by_name_and_port_and_protocol(self, service_id, container_port, domain_name, protocol,
-                                                 domain_path=None):
+    def get_domain_by_name_and_port_and_protocol(self, service_id, container_port, domain_name, protocol, domain_path=None):
         if domain_path:
             try:
                 return ServiceDomain.objects.get(
@@ -574,8 +584,7 @@ class ServiceDomainRepository(object):
 
     def get_domain_by_name_and_path_and_protocol(self, domain_name, domain_path, protocol):
         if domain_path:
-            return ServiceDomain.objects.filter(domain_name=domain_name, domain_path=domain_path,
-                                                protocol=protocol).all()
+            return ServiceDomain.objects.filter(domain_name=domain_name, domain_path=domain_path, protocol=protocol).all()
         else:
             return None
 
@@ -643,8 +652,7 @@ class ServiceDomainRepository(object):
     def get_service_domains(self, service_id):
         return ServiceDomain.objects.filter(service_id=service_id).all()
 
-    def create_service_domains(self, service_id, service_name, domain_name, create_time, container_port, protocol,
-                               http_rule_id,
+    def create_service_domains(self, service_id, service_name, domain_name, create_time, container_port, protocol, http_rule_id,
                                tenant_id, service_alias, region_id):
         ServiceDomain.objects.create(
             service_id=service_id,
@@ -803,8 +811,7 @@ class ServiceTcpDomainRepository(object):
 
     def get_tcpdomain_by_name_and_port(self, service_id, container_port, end_point):
         try:
-            return ServiceTcpDomain.objects.get(service_id=service_id, container_port=container_port,
-                                                end_point=end_point)
+            return ServiceTcpDomain.objects.get(service_id=service_id, container_port=container_port, end_point=end_point)
         except ServiceTcpDomain.DoesNotExist:
             return None
 
@@ -816,8 +823,7 @@ class ServiceTcpDomainRepository(object):
                     query = Q(region_id=region_id, end_point__icontains=":{}".format(hostport[1]))
                     return ServiceTcpDomain.objects.filter(query)
                 query_default_endpoint = "0.0.0.0:{0}".format(hostport[1])
-                query = Q(region_id=region_id, end_point=end_point) | Q(region_id=region_id,
-                                                                        end_point=query_default_endpoint)
+                query = Q(region_id=region_id, end_point=end_point) | Q(region_id=region_id, end_point=query_default_endpoint)
                 return ServiceTcpDomain.objects.filter(query)
             return None
         except ServiceTcpDomain.DoesNotExist:
@@ -906,6 +912,7 @@ port_repo = TenantServicePortRepository()
 image_service_relation_repo = ImageServiceRelationRepository()
 domain_repo = ServiceDomainRepository()
 volume_repo = TenantServiceVolumnRepository()
+config_file_repo = ComponentConfigurationFileRepository()
 mnt_repo = TenantServiceMntRelationRepository()
 dep_relation_repo = TenantServiceRelationRepository()
 extend_repo = ServiceExtendRepository()
